feature: Cryptocurrency Trading Enhancements

+++ server/src/main/java/mg/exchange/models/User.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 import java.math.BigDecimal;
 
@@ -18,7 +17,7 @@ public class User {
     @Id
     @Column(name = "id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "fiat_balance", nullable = false, precision = 18, scale = 2)
     private BigDecimal fiatBalance;
